requirements.txt
greenlet==3.2.4
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
setuptools==80.9.0
typing_extensions==4.15.0
wheel==0.45.1
